feat: implement bytecode emulator backend
The native code generation backend is held up on a couple issues that we
don't want to spend time at the moment attempting to address - namely
that lowering to WebAssembly presents some difficult issues around how
to support some of what our native backend wants or needs to do in order
to bring all the pieces of the compiler and runtime together. The lack
of any stack switching intrinsics native to Wasm is a major obstacle
here.

As a result, what this commit does is extend the existing compiler to
support generation of our own bytecode, executed by an emulator which is
linked in to any generated executables. From the perspective of the end
user, there is no noticeable difference between native compilation and
this backend, however there are a few constraints (addressable) that are
not handled here - namely compilation of static/dynamic libraries, and
linking modules previously compiled. This can be handled by performing
our linking process, but that hasn't been implemented and currently
isn't important.

A significant amount of runtime infrastructure was also implemented
during this time, including signal management, system services, and a
rework of the term encoding scheme to support some additional types. The
garbage collector was implemented and is integrated with the emulator.

The next outstanding bit of work is aggressively testing what has been
built here, and expanding upon it in a few ways:

* Build on the current primitives to implement the init procedure
* Implement ETS, persistent term, and atomics services
* Implement the standard library on top of the above

+++ compiler/syntax_base/src/literals.rs
@@ -6,7 +6,7 @@ use std::hash::{Hash, Hasher};
 use firefly_binary::BitVec;
 use firefly_diagnostics::{SourceSpan, Spanned};
 use firefly_intern::Symbol;
-use firefly_number::{Float, Integer};
+use firefly_number::{Float, Int};
 
 use crate::*;
 
@@ -27,7 +27,7 @@ impl Literal {
         }
     }
 
-    pub fn integer<I: Into<Integer>>(span: SourceSpan, i: I) -> Self {
+    pub fn integer<I: Into<Int>>(span: SourceSpan, i: I) -> Self {
         Self {
             span,
             annotations: Annotations::default(),
@@ -90,7 +90,7 @@ impl Literal {
         }
     }
 
-    pub fn as_integer(&self) -> Option<&Integer> {
+    pub fn as_integer(&self) -> Option<&Int> {
         match &self.value {
             Lit::Integer(ref i) => Some(i),
             _ => None,
@@ -133,7 +133,7 @@ impl Ord for Literal {
 #[derive(Clone, PartialEq, Eq)]
 pub enum Lit {
     Atom(Symbol),
-    Integer(Integer),
+    Integer(Int),
     Float(Float),
     Nil,
     Cons(Box<Literal>, Box<Literal>),
